+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 #### 1.9.8-SNAPSHOT
 
+- **[UPDATE]** Updated to Kotlin 1.2.51.
+- **[UPDATE]** Updated to Kotlin Coroutines 0.24.0.
+
 #### 1.9.8-b3
 
 - **[UPDATE]** Updated to Kotlin 1.2.41.